etc-keeper for Gentoo
Okay, here it is in a somewhat cleaner form. As far as the idea of using git
branches to do installation control in /etc, I was unable to figure out a
consistently safe/useful/sensible spot to branch at, so I'm shelving the idea
for now. I have included a changelog in the tarball, and the "ebuilds"
directory contains everything neccessary to be a paludis-valid repository.