remove versioned dependencies on bzr 1.4
I was told that bzr 1.4 would be released last weekend, but that does not
seem to be true, (unless they meant that 1.4rc1 would be released). I don't
want to keep a RC bug open blocked on a bzr release. The versioned deps
arn't really needed; building with an older bzr will drop the bzr hook
and otherwise work.